rename old likert summary
