+++ src/layout/Likert/Summary/LikertSummaryComponent.module.css
